+++ cost-optimization/hpa-config-recommender/buildtest.dockerfile
