+++ cost-optimization/hpa-config-recommender/deploy/variables.tf
