+++ cost-optimization/hpa-config-recommender/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "workloadrecommender"
+name = "hpaconfigrecommender"
 version = "0.0.41"
 authors = [
     { name="Google LLC", email="ameenahb@google.com" }
@@ -13,7 +13,6 @@ description = "Workload Recommender Core Library"
 readme = "README.md"
 requires-python = ">=3.10"
 
-license = {text = "The Unlicense"}
 
 classifiers = [
     "Programming Language :: Python :: 3",
@@ -32,3 +31,4 @@ dependencies = [
     "google-cloud-asset==3.27.0",
     "google_cloud_bigquery==3.29.0"
 ]
+license = {text = "The Unlicense"}
