+++ cost-optimization/hpa-config-recommender/requirements.in
