+++ cost-optimization/hpa-config-recommender/requirements.txt
