+++ cost-optimization/hpa-config-recommender/setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = workloadrecommender
+name = hpaconfigrecommender
 version = 0.0.38
 
 [options]
