+++ cost-optimization/hpa-config-recommender/src/hpaconfigrecommender.egg-info/requires.txt
@@ -1,0 +1,9 @@
+pandas==2.2.2
+pyarrow<19.0.0
+numpy==1.26.4
+aiohttp>=3.8.0
+asyncio==3.4.3
+requests==2.32.3
+google-api-core==2.22.0
+google-cloud-asset==3.27.0
+google_cloud_bigquery==3.29.0
