+++ cost-optimization/hpa-config-recommender/src/hpaconfigrecommender.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+hpaconfigrecommender
