+++ cost-optimization/hpa-config-recommender/src/hpaconfigrecommender/utils/models.py
