+++ cost-optimization/hpa-config-recommender/tests/test_run_workload_simulation_plan.py
@@ -17,10 +17,10 @@
 import unittest
 import pandas as pd
 import numpy as np
-from workloadrecommender.utils.models import (
+from hpaconfigrecommender.utils.models import (
     WorkloadDetails
 )
-from workloadrecommender.run_workload_simulation_plan import (
+from hpaconfigrecommender.run_workload_simulation_plan import (
     get_simulation_plans,
     _is_workload_balanced,
     _calculate_recommended_max_cpu_capacity,
@@ -29,7 +29,7 @@ from workloadrecommender.run_workload_simulation_plan import (
     convert_data_types,
     _get_proposed_memory_recommendation
 )
-from workloadrecommender.utils.config import Config
+from hpaconfigrecommender.utils.config import Config
 
 class TestHPASimulationPlan(unittest.TestCase):
     """Unit tests for HPA simulation algorithms."""
